requirements.txt
geopy==2.4.1
requests==2.32.3
secure-smtplib==0.1.1
python-dotenv==1.0.1
